backend/requirements.txt
fastapi
uvicorn
neo4j
langchain
langchain-openai
pydantic
scikit-learn
networkx
python-dotenv